Use markers instead of pilot's suite facility for test selection
suite causes non-selected tests to be skipped, which means that in our
two pytest invocations, non-selected tests show up as skipped rather
than deselected (and hence absent in a results file).

As a result the tools that process results file get confused and, at
best, display tests as both being skipped and successful.

This approach uses the built in mark facility and configures a default
mark selector that deselects container tests.