Fix build
I did an incomplete job of porting in the improvements from
ngfw-edl-server. The project name is now parametized (which would have
prevented this in the first place).

Because we're now mounting the project directory into the build
container as a temporary/overlay directory, changes do not persist
between RUN directives. Mount a temporary directory into the `dist`
directory so that the wheel built by one RUN directive can be installed
by another.